+++ src/rules/custom-property-empty-line-before/index.js
@@ -1,4 +1,5 @@
 import {
+  hasEmptyLine,
   isCustomProperty,
   optionsHaveException,
   report,
@@ -62,12 +63,10 @@ export default function (expectation, options) {
           expectEmptyLineBefore = !expectEmptyLineBefore
         }
 
-        const before = decl.raws["before"]
-        const emptyLineBefore = before.indexOf("\n\n") !== -1
-          || before.indexOf("\n\r\n") !== -1
+        const hasEmptyLineBefore = hasEmptyLine(decl.raws["before"])
 
         // Return if the expectation is met
-        if (expectEmptyLineBefore === emptyLineBefore) { return }
+        if (expectEmptyLineBefore === hasEmptyLineBefore) { return }
 
         const message = expectEmptyLineBefore ? messages.expected : messages.rejected
         report({